+++ tests/fixtures/composition-test.vue
@@ -1,0 +1,18 @@
+<template>
+	<code>String: "{{ props.someString }}"
+		Int: {{ props.someInt }}
+	</code>
+</template>
+
+<script setup>
+const props = defineProps({
+	someString: {
+		type: String,
+		required: true,
+	},
+	someInt: {
+		type: Number,
+		required: true,
+	},
+})
+</script>